Updated to upgrade and to add link to map to open in new window

+++ tools/createmap.py
@@ -7,5 +7,8 @@ if __name__=="__main__":
     path_to_mojomap="."
     if not os.path.exists(path):
         os.mkdir(path)
-    os.system("cp -arv %s %s/mojomaps" %(os.path.join(path_to_mojomap,"web"),path))
+    if os.path.exists(os.path.join(path,"mojomaps")):
+    	os.system("cp -arv %s/* %s/mojomaps" %(os.path.join(path_to_mojomap,"web"),path))
+    else:
+    	os.system("cp -arv %s %s/mojomaps" %(os.path.join(path_to_mojomap,"web"),path))
